requirements.txt
Flask>=2.2
flask_sqlalchemy>=3.0
flask_migrate>=4.0
flask_jwt_extended>=4.4
flask_cors>=4.0
python-dotenv>=1.0
gunicorn>=20.1
celery>=5.2
redis>=4.0
psycopg2-binary>=2.9
openai>=1.0
Werkzeug>=2.2
